Fixed maxiEnv::adsr
When sustain was set to 0, maxiEnv::adsr would not function properly. With the new changes, it is now simple to create pluck sounds by setting the sustain to 0, and then adjusting decay and release. When the trigger is released, the envelope will automatically move to the next phase, allowing for rapid playing of plucked notes as well